Move files and empty folder's __init__
Move files to another folder, and delete empty folder's __init__ if the folder was empty
